Move sting module into components directory
