+++ src/components/sting-component/sting-component.module.css
