+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import { StingComponent } from './sting.tsx'
+import { StingComponent } from './components/sting-component/sting-component'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
